refactor(live-tokens): split token data types
Move live-token data types into a leaf module so perps formatting and Discover market cards can depend on the token shape without importing the live token store itself.

The store re-exports the same types for existing callers, so this only changes the dependency graph. It gives later perps and chart commits a cycle-safe type boundary from the start.
